Betsy/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
paho-mqtt==1.6.1
python-dotenv==1.0.1
bcrypt==4.1.3